requirements.txt
fastapi
uvicorn
jinja2
sqlalchemy
stripe
python-dotenv
passlib[bcrypt]
python-jose
requests
